# server/swagger.yaml
openapi: 3.0.0
info:
  title: MAP API
  version: 1.1.0
  description: |
    MAP API provides endpoints for managing users, authentication, and more.
    It supports user registration, sign-in, password management, role selection, and 2FA...
  contact:
    name: MAP
    email: hunter@pentabell.fr
servers:
  - url: http://localhost:4000/api/v1
  - url: http://192.168.10.137:4000/api/v1
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
...
paths:
...
  /interviews:
    post:
      tags:
        - Interviews
      summary: Schedule a new interview (Recruiter only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [applicationId, date, durationMinutes]
              properties:
                applicationId:
                  type: string
                  format: uuid
                  example: "b3f8c123-4567-89ab-cdef-1234567890ab"
                date:
                  type: string
                  format: date-time
                  example: "2025-09-12T14:00:00Z"
                durationMinutes:
                  type: integer
                  example: 60
                notes:
                  type: string
                  example: "Technical interview with recruiter"
                type:
                  type: string
                  enum: [VOICE, VIDEO]
                  example: "VIDEO"
                status:
                  type: string
                  enum: [SCHEDULED, COMPLETED, CANCELLED]
                  example: "SCHEDULED"
      responses:
        201:
          description: Interview scheduled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
        400:
          description: Missing or invalid fields
        401:
          description: Unauthorized

    get:
      tags:
        - Interviews
      summary: Get all interviews of recruiter (Recruiter only)
      security:
        - cookieAuth: []
      responses:
        200:
          description: List of recruiter interviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interview"
        401:
          description: Unauthorized
  /interviews/{id}:
    get:
      tags:
        - Interviews
      summary: Get interview by ID (Recruiter only)
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Interview found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
        404:
          description: Interview not found

    patch:
      tags:
        - Interviews
      summary: Update interview (Recruiter only)
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date-time
                durationMinutes:
                  type: integer
                notes:
                  type: string
                type:
                  type: string
                  enum: [VOICE, VIDEO]
                status:
                  type: string
                  enum: [SCHEDULED, COMPLETED, CANCELLED]
      responses:
        200:
          description: Interview updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
        404:
          description: Interview not found

    delete:
      tags:
        - Interviews
      summary: Cancel interview (Recruiter only)
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Interview cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Interview cancelled successfully"
        404:
          description: Interview not found
...
components:
  schemas:
    Interview:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "c7d8e123-4567-89ab-cdef-1234567890ab"
        applicationId:
          type: string
          format: uuid
          example: "b3f8c123-4567-89ab-cdef-1234567890ab"
        recruiterId:
          type: string
          format: uuid
          example: "a1b2c3d4-5678-90ef-1234-567890abcdef"
        candidateId:
          type: string
          format: uuid
          example: "f9e8d7c6-1234-5678-90ab-cdef12345678"
        date:
          type: string
          format: date-time
          example: "2025-09-17T14:00:00"
        durationMinutes:
          type: integer
          example: 60
        notes:
          type: string
          example: "Technical interview with recruiter"
        type:
          type: string
          enum: [voice, video]
          example: "video"
        status:
          type: string
          enum: [scheduled, completed, cancelled, missed]
          example: "scheduled"